+++ clients/benchmarks/client.cpp
@@ -294,13 +294,6 @@ try
         //     "                           The order in which a series of transformations are applied.\n"
         //     "                           ")
 
-        // ("evect",
-        //  value<char>()->default_value('N'),
-        //     "N = none, V = compute eigenvectors of the matrix,\n"
-        //     "                           I = compute eigenvectors of the tridiagonal matrix.\n"
-        //     "                           Indicates how the eigenvectors are to be calculated and stored.\n"
-        //     "                           ")
-
         // ("fast_alg",
         //  value<char>()->default_value('O'),
         //     "O = out-of-place, I = in-place.\n"
@@ -312,11 +305,17 @@ try
         //     "Increment between values in vector x.\n"
         //     "                           ")
 
-        // ("itype",
-        //  value<char>()->default_value('1'),
-        //     "1 = Ax, 2 = ABx, 3 = BAx.\n"
-        //     "                           Problem type for generalized eigenproblems.\n"
-        //     "                           ")
+        ("itype",
+         value<char>()->default_value('1'),
+            "1 = Ax, 2 = ABx, 3 = BAx.\n"
+            "                           Problem type for generalized eigenproblems.\n"
+            "                           ")
+
+        ("jobz",
+         value<char>()->default_value('N'),
+            "N = none, V = compute eigenvectors of the matrix,\n"
+            "                           Indicates how the eigenvectors are to be calculated and stored.\n"
+            "                           ")
 
         ("side",
          value<char>(),
@@ -375,8 +374,8 @@ try
     argus.validate_svect("jobu");
     argus.validate_svect("jobv");
     // argus.validate_workmode("fast_alg");
-    // argus.validate_evect("evect");
-    // argus.validate_itype("itype");
+    argus.validate_itype("itype");
+    argus.validate_evect("jobz");
 
     // select and dispatch function test/benchmark
     hipsolver_dispatcher::invoke(function, precision, argus);